+++ src/begin/userscript_definition.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name	Twitter - Mark Last Read
-// @version 1.7.0
+// @version 1.8.0
 // @grant   none
 // @include https://*twitter.com/*
 // @include https://*x.com/*